// src/app/Pages/Models/emailtemplate.ts
export class emailtemplate {
    ID: number;
    TEMPLATE_NAME: string = '';
    SUBJECT: string = '';
    BODY: any;
    ATTACHMENTS: any;
    IS_ACTIVE: boolean = true;
    BODY_VALUES: any;
    BODY_TEXT: any;
    TEMPLATE_BODY: any;
    DESCRIPTION: string = '';
    LANGUAGE_CODE: any;
    TEMPLATE_CATEGORY_ID: any;
}